+++ Changelog.markdown
@@ -5,6 +5,7 @@
 #### Version 0.2.3 (2013-06-29):
 
 * __new__:  `required` optional argument in `clap.Parser.add()`, if passed and option is not found in input when `check()`ing error is raised,
+* __new__:  `gethint()` method in `clap.Parser`,
 
 
 ----